split plotting functions into separate files and module
